QAC_line_bot/requirements.txt
line-bot-sdk
flask
oauth2client
gspread
datetime
